requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
black==24.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.13
django-environ==0.11.2
djangorestframework==3.15.2
idna==3.7
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
requests==2.32.3
sqlparse==0.5.0
tomli==2.0.1
typing-extensions==4.12.2
urllib3==2.2.2
psycopg2-binary>=2.8.6
